Remove unused code, move codes to proper modules.
