+++ src/crypto/mod.rs
@@ -1,11 +1,9 @@
 pub mod private_key;
 pub mod public_key;
 pub mod schnorr;
-pub mod util;
 pub mod zil_address;
 
 pub use k256::ecdsa::Signature;
 pub use private_key::PrivateKey;
 pub use public_key::PublicKey;
-pub use util::*;
 pub use zil_address::ZilAddress;
